fixup! [#1163] added TDrive Iframe loading
