+++ __test__/features/Tdrive/tdriveUrlUtils.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 
-import { resolveTdriveUrl, isTdriveEnabled } from '@common/utils/tdriveUrlUtils'
+import { resolveTdriveUrl } from '@common/utils/tdriveUrlUtils'
 
 describe('resolveTdriveUrl', () => {
   const originalUrl = window.TDRIVE_INTENT_URL
@@ -55,41 +55,5 @@ describe('resolveTdriveUrl', () => {
   it('should leave unknown expressions untouched', () => {
     window.TDRIVE_INTENT_URL = 'https://drive.example.com/{unknown}'
     expect(resolveTdriveUrl()).toBe('https://drive.example.com/{unknown}')
-  })
-})
-
-describe('isTdriveEnabled', () => {
-  const originalEnabled = window.TDRIVE_ENABLED
-  const originalUrl = window.TDRIVE_INTENT_URL
-
-  afterEach(() => {
-    window.TDRIVE_ENABLED = originalEnabled
-    window.TDRIVE_INTENT_URL = originalUrl
-  })
-
-  it('returns false when TDRIVE_ENABLED is undefined', () => {
-    // @ts-expect-error simulate missing configuration
-    window.TDRIVE_ENABLED = undefined
-    window.TDRIVE_INTENT_URL = 'https://drive.example.com'
-    expect(isTdriveEnabled()).toBe(false)
-  })
-
-  it('returns false when TDRIVE_ENABLED is false', () => {
-    window.TDRIVE_ENABLED = false
-    window.TDRIVE_INTENT_URL = 'https://drive.example.com'
-    expect(isTdriveEnabled()).toBe(false)
-  })
-
-  it('returns false when TDRIVE_INTENT_URL is not set', () => {
-    window.TDRIVE_ENABLED = true
-    // @ts-expect-error simulate missing configuration
-    window.TDRIVE_INTENT_URL = undefined
-    expect(isTdriveEnabled()).toBe(false)
-  })
-
-  it('returns true when both TDRIVE_ENABLED and TDRIVE_INTENT_URL are set', () => {
-    window.TDRIVE_ENABLED = true
-    window.TDRIVE_INTENT_URL = 'https://drive.example.com'
-    expect(isTdriveEnabled()).toBe(true)
   })
 })
